requirements.txt
PyPDF2
python-docx
python-pptx
